requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<3.0
scipy>=1.10,<2.0
statsmodels>=0.14,<1.0
matplotlib>=3.7,<4.0
scikit-learn>=1.4,<2.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0
holidays>=0.50
ruptures>=1.1
streamlit>=1.35,<2.0
torch>=2.0,<3.0
